--- frontend/src/styles/Variables.ts
@@ -1,77 +1,0 @@
-export const Variables = {
-  colors: {
-    primary: '#fe4902',
-    surface_default: '#F8F8F8',
-    surface_white: '#FFFFFF',
-    surface_transparent_white_35: 'rgba(255, 255, 255, 0.35)',
-    surface_alt: '#C4C4C4',
-    surface_strong: '#434343',
-    surface_transparent_darkgrey_75: '#232323BF',
-    surface_black: '#000000',
-    surface_weak: '#EFEFEF',
-    surface_transparent_black_50: 'rgba(0, 0, 0, 0.5)', //50% 투명도
-    surface_point: '#FF8A8A',
-    text_default: '#000000',
-    text_alt: '#656565',
-    text_weak: '#777777',
-    text_white: '#FFFFFF',
-    text_transparent_white_70: 'rgba(255, 255, 255, 0.7)',
-    text_point: '#0051FF',
-    player_blue: '#61B8FF',
-    player_grey: '#E3E3E3',
-    player_red: '#FF7275',
-    player_green: '#71F084',
-    player_orange: '#FFAD55',
-    player_purple: '#9B4DFF',
-    player_yellow: '#FFEB3B',
-    player_pink: '#FF61C7',
-    player_cyan: '#00BCD4',
-    player_brown: '#795548',
-    surface_word_default: '#D9D9D9',
-    surface_word_weak: '#EAFFF3',
-    surface_word_medium: '#DAFFEB',
-    surface_word_strong: '#B6FAD4',
-    text_word_weak: '#62CA90',
-    text_word_medium: '#00A549',
-    text_word_strong: '#00602B',
-    surface_youtube_weak: '#FFEBEB',
-    surface_file_weak: '#DCEBFF',
-    surface_green_strong: '#35BD79',
-    surface_green_weak: '#67E7A5',
-    surface_orange_strong: '#FF8548',
-    surface_oragne_weak: '#FFB184'
-  },
-  typography: {
-    font_light_14: '300 14px "Pretendard", sans-serif',
-    font_medium_48: '500 48px "Pretendard", sans-serif',
-    font_medium_36: '500 36px "Pretendard", sans-serif',
-    font_medium_24: '500 24px "Pretendard", sans-serif',
-    font_medium_20: '500 20px "Pretendard", sans-serif',
-    font_medium_18: '500 18px "Pretendard", sans-serif',
-    font_medium_16: '500 16px "Pretendard", sans-serif',
-    font_medium_14: '500 14px "Pretendard", sans-serif',
-    font_bold_72: '700 72px "Pretendard", sans-serif',
-    font_bold_56: '700 56px "Pretendard", sans-serif',
-    font_bold_48: '700 48px "Pretendard", sans-serif',
-    font_bold_36: '700 36px "Pretendard", sans-serif',
-    font_bold_32: '700 32px "Pretendard", sans-serif',
-    font_bold_28: '700 28px "Pretendard", sans-serif',
-    font_bold_24: '700 24px "Pretendard", sans-serif',
-    font_bold_20: '700 20px "Pretendard", sans-serif',
-    font_bold_18: '700 18px "Pretendard", sans-serif',
-    font_bold_16: '700 16px "Pretendard", sans-serif'
-  },
-  shadow: {
-    shadow_floating: '0px 4px 8px #00000040'
-  },
-  spacing: {
-    spacing_xs: '4px',
-    spacing_sm: '8px',
-    spacing_md: '16px',
-    spacing_lg: '24px',
-    spacing_xl: '32px',
-    spacing_xxl: '48px',
-    spacing_3xl: '64px',
-    spacing_4xl: '80px'
-  }
-};
